requirements.txt
Flask==3.0.0
gunicorn==21.2.0
psycopg2-binary
python-dotenv